added test_horse_relationships_create to horse_unit_test.py, removed horse_relationships_unit_test.py
